Working on the search bar for the job postings. Trying different things out, will start working on implementing a job router
