+++ package.json
@@ -8,7 +8,8 @@
     "dev": "next dev",
     "postinstall": "prisma generate",
     "lint": "next lint",
-    "start": "next start"
+    "start": "next start",
+    "ts-node": "ts-node --compiler-options '{\"module\":\"CommonJS\"}'"
   },
   "dependencies": {
     "@clerk/nextjs": "^4.25.1",
@@ -25,20 +26,21 @@
     "@trpc/react-query": "^10.37.1",
     "@trpc/server": "^10.37.1",
     "flowbite": "^1.8.1",
-    "framer-motion": "^10.16.4",
     "flowbite-react": "^0.6.4",
+    "framer-motion": "^10.16.4",
     "next": "^13.4.19",
     "preline": "^1.9.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "superjson": "^1.13.1",
     "swiper": "^10.3.1",
+    "swr": "^2.2.4",
     "zod": "^3.21.4"
   },
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.10",
     "@types/eslint": "^8.44.2",
-    "@types/node": "^18.16.0",
+    "@types/node": "^18.18.6",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.3.0",
@@ -51,7 +53,11 @@
     "prettier-plugin-tailwindcss": "^0.5.1",
     "prisma": "^5.1.1",
     "tailwindcss": "^3.3.3",
-    "typescript": "^5.1.6"
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
+  },
+  "prisma": {
+    "seed": "ts-node --compiler-options {\"module\":\"CommonJS\"} prisma/seed.ts"
   },
   "ct3aMetadata": {
     "initVersion": "7.20.2"
